Reformat tools for handling Tristan data
Move more data handling tools from tristan to tristan.data.
